Patched build loading to reset all selections before loading the build to ensure there aren't any stray effects carried over.

+++ Buildframe/Methods/Calculation/Weapon.cs
@@ -139,7 +139,7 @@ namespace Buildframe.Methods.Calculation
             WriteLineIfDebug("Rolling enervate");
 
             double critChance = calculateModCritChance(stats);
-            int steps = 0;
+            double steps = 0;
             Random rnd = new Random();
             int degreeOfAccuracy = 100;
             for (int i = 0; i <= degreeOfAccuracy; i++)
